pyproject.toml
[project]
name = "GenEC"
version = "0.1.0"
description = "Generic Extraction & Comparison Tool"
readme = "README.md"
readme-content-type = "text/markdown"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
  { name = "Remy Kroese", email = "remykroese@hotmail.com" }
]
keywords = ["data", "data-analysis", "data-processing", "extraction", "comparison", "automation", "cli"]

dependencies = [
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
]

[dependency-groups]
dev = [
    "autopep8>=2.3.2",
    "flake8>=7.3.0",
    "flake8-docstrings>=1.7.0",
    "flake8-pyproject>=1.2.3",
    "pycodestyle>=2.14.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-repeat>=0.9.4",
]
dist = [
    "pyinstaller>=6.15.0"
]

[project.urls]
Homepage = "https://github.com/remykroese/GenEC"
Docs = "https://github.com/RemyKroese/GenEC/tree/main/docs/"
Issue_tracker = "https://github.com/remykroese/GenEC/issues"

[tool.pytest.ini_options]
addopts = "-ra --strict-markers"
markers = [
    "unit",
    "system"
]
filterwarnings = [
    "ignore::SyntaxWarning:prettytable"
]

[tool.flake8]
property-decorators="true"
exclude = [
  '.venv/*',
  '*prettytable*',
  '*__init__.py*',
  'tests/*'
]
max-line-length = 150
count = true
statistic = true
docstring-convention = "numpy"
ignore-self-only-init = true

[tool.coverage.run]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py"
]

[tool.coverage.report]
show_missing = true
skip_covered = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:"
]

[tool.uv.sources]
pypi = { url = "https://pypi.org/simple" }
